fix: delete drafts
