+++ image-galery/css/style.css
@@ -271,6 +271,7 @@ body {
   left: 50%;
   transform: translate(-50%, -50%);
   opacity: 0.1;
+  cursor: default;
 }
 
 .header-container {
@@ -332,7 +333,7 @@ body {
   opacity: 0;
   width: 22px;
   height: 50px;
-  background-image: url(../assets/svg/Delete-80_icon-icons.com_57340.png); /* замените путь на изображение крестика */
+  background-image: url(../assets/svg/Delete-80_icon-icons.com_57340.png); /*изображение крестика */
   background-size: contain;
   background-repeat: no-repeat;
   background-position: center center;
