+++ image-galery/index.html
@@ -35,47 +35,9 @@
 				</form>
 			</div>
 
-
-			<!-- <form class="header-form">
-				<div class="header-div">
-					<input class="header-input header-input__text" type="text" placeholder="Search..." autofocus autocomplete="off">
-					<span class="header-input__clear"></span>
-				</div>
-
-				<input class="header-input header-input__submit" type="submit" name="input-submit" value="">
-			</form> -->
 		</header>
 
 		<main class="main">
-
-			<!-- <div class="cover">
-				<div class="cover-inner cover-inner--ratio">
-						<div class="cover-content cover-content--01">
-
-						</div>
-				</div>
-			</div>
-			<div class="cover">
-				<div class="cover-inner cover-inner--ratio">
-						<div class="cover-content cover-content--02">
-
-						</div>
-				</div>
-			</div>
-			<div class="cover">
-				<div class="cover-inner cover-inner--ratio">
-						<div class="cover-content cover-content--03">
-
-						</div>
-				</div>
-			</div>
-			<div class="cover">
-				<div class="cover-inner cover-inner--ratio">
-						<div class="cover-content cover-content--04">
-
-						</div>
-				</div>
-			</div> -->
 
 		</main>
 
@@ -97,7 +59,6 @@
 		</footer>
 
 	</div>
-	<!-- <script type="module" src="./js/index03.js"></script> -->
-	<script type="module" src="./js/index04.js"></script>
+	<script type="module" src="./js/index.js"></script>
 </body>
 </html>
